+++ R/blockAverage.R
@@ -62,10 +62,10 @@
 #' @export
 #'
 #' @examples
-#' t_dat <- make_whole(sst_Med)
-#' res <- detect(t_dat, climatology_start = 1983, climatology_end = 2012) # using default values
-#' out <- block_average(res)
-#' summary(glm(count ~ year, out, family = "poisson"))
+#' # t_dat <- make_whole(sst_Med)
+#' # res <- detect(t_dat, climatology_start = 1983, climatology_end = 2012) # using default values
+#' # out <- block_average(res)
+#' # summary(glm(count ~ year, out, family = "poisson"))
 #'
 #' \dontrun{
 #' plot(out$year, out$count, col = "salmon", pch = 16,