+++ R/makeWhole.R
@@ -3,11 +3,17 @@
 #' Takes a series of dates and temperatures, and if irregular (but ordered), inserts
 #' missing dates and fills correpsonding temperatures with NAs.
 #'
-#' @param data A data frame with columns headed \code{t} and \code{temp} for
-#' date and temperature data, respectively. Ordered daily data are expected, and
-#' although missing values (NA) can be accommodated, the function is only
-#' recommended when NAs occur infrequently, preferably at no more than 3
-#' consequtive days.
+#' @param data A data frame with columns for date and temperature data.
+#' Ordered daily data are expected, and although missing values (NA) can be
+#' accommodated, the function is only recommended when NAs occur infrequently,
+#' preferably at no more than 3 consecutive days.
+#' @param x A column with the daily time vector (see details). For backwards
+#' compatibility, the column is named \code{t} by default.
+#' @param y A column with the response vector. RmarineHeatWaves version <= 0.15.9
+#' assumed that this would be daily seawater temperatures, but as of version 0.16.0
+#' it may be any arbitrary measurement taken at a daily frequency. The default
+#' remains temperature, and the default column name is therefore \code{temp}, again
+#' hopefully ensuring backwards compatibility.
 #'
 #' @details
 #' Upon import, the package uses `zoo` and `lubridate` to process the input
@@ -38,32 +44,47 @@
 #' headed \code{doy} (day-of-year) is the Julian day running from 1 to 366, but
 #' modified so that the day-of-year series for non-leap-years runs 1...59 and
 #' then 61...366. For leap years the 60th day is February 29. See the example,
-#' below. The \code{date} column is a series of dates of class \code{Date},
-#' while \code{temp} is temperature. This time series will be uninterrupted and
-#' continuous daily values between the first and last dates of the input data.
+#' below. The other two columns take the names of \code{x} and \code{y}, if supplied,
+#' or it will be \code{t} and \code{temp} in case the default values were used.
+#' The \code{x} (or \code{t}) column is a series of dates of class \code{Date},
+#' while \code{y} (or \code{temp}) is the measured variable. This time series will
+#' be uninterrupted and continuous daily values between the first and last dates
+#' of the input data.
 #' @export
 #'
 #' @author Smit, A. J.
 #'
 #' @examples
 #' require(dplyr); require(tidyr); require(lubridate)
-#' ts_dat <- make_whole(sst_WA)
+#' ts_dat <- make_whole(sst_WA) # default columns "t" and "temp", in that order
 #' clim_start <- "1983-01-01"
 #' clim_end <- "2012-12-31"
 #' ts_dat %>%
-#' filter(date >= clim_start & date <= clim_end) %>%
-#'   mutate(date = year(date)) %>%
-#'   spread(date, temp) %>%
+#' filter(t >= clim_start & t <= clim_end) %>%
+#'   mutate(t = year(t)) %>%
+#'   spread(t, temp) %>%
 #'   filter(doy >= 55 & doy <= 65)
-make_whole <- function(data) {
-  temp <- NULL
-  data <- data %>%
-    dplyr::group_by(t) %>%
-    dplyr::summarise(temp = mean(temp, na.rm = TRUE))
-  tSeries <- zoo::zoo(data$temp, data$t)
+make_whole <- function(data, x, y) {
+  if (missing(x)) {
+    x <- "t"
+  } else {
+    x <- x
+  }
+  if (missing(y)) {
+    y <- "temp"
+  } else {
+    y <- y
+  }
+  dat <- data.frame(x = data[, x],
+                    y = data[, y])
+  dat <- dat %>%
+    dplyr::group_by(x) %>%
+    dplyr::summarise(y = mean(y, na.rm = TRUE)) %>%
+    dplyr::ungroup()
+  tSeries <- zoo::zoo(dat$y, dat$x)
   ser <-
-    data.frame(t = seq(stats::start(tSeries), stats::end(tSeries), by = "1 day"))
-  ser <- zoo::zoo(rep(NA, length(ser$t)), order.by = ser$t)
+    data.frame(x = seq(stats::start(tSeries), stats::end(tSeries), by = "1 day"))
+  ser <- zoo::zoo(rep(NA, length(ser$x)), order.by = ser$x)
   tSeries <- merge(ser, tSeries)[, 2]
 
   feb28 <- 59
@@ -72,7 +93,7 @@ make_whole <- function(data) {
     data.frame(
       doy = lubridate::yday(tSeries),
       date = as.Date(as.POSIXct(tSeries)),
-      temp = tSeries,
+      y = tSeries,
       row.names = NULL ###
     ) %>%
     dplyr::mutate(doy = ifelse(
@@ -80,20 +101,7 @@ make_whole <- function(data) {
       ifelse(doy > feb28, doy + 1, doy),
       doy
     ))
-
-  # property <- c("start date",
-  #               "end date",
-  #               "no. NAs")
-  #               # "% NAs",
-  #               # "mean no. NAs/year",
-  #               # "min. no. NAs/year",
-  #               # "max. no. NAs/year")
-  # value <- c(as.Date(tSeries$date[1]),
-  #            as.Date(tSeries$date[length(tSeries$date)]),
-  #            sum(is.na(tSeries$temp))
-  #            )
-  # print(kable(cbind(property, value)))
-  # print(paste("start date: ", tSeries$date[1]))
-  # print(paste("end date: ", tSeries$date[length(tSeries$date)]))
-  # return(tSeries)
+  colnames(tSeries) <- c("doy", x, y)
+  return(tSeries)
 }
+